+++ objgraph.py
@@ -29,7 +29,7 @@ Released under the MIT licence.
 __author__ = "Marius Gedminas (marius@gedmin.as)"
 __copyright__ = "Copyright (c) 2008-2015 Marius Gedminas and contributors"
 __license__ = "MIT"
-__version__ = "1.9.0.dev0"
+__version__ = "2.0.0.dev0"
 __date__ = "2014-05-15"
 
 
@@ -44,12 +44,16 @@ import subprocess
 import tempfile
 import sys
 import itertools
+import logging
+
+LOG = logging.getLogger('objgraph')
 
 
 try:
     basestring
 except NameError:
     # Python 3.x compatibility
+    # pylint: disable=W0622
     basestring = str
 
 try:
@@ -84,7 +88,7 @@ def count(typename, objects=None):
     if objects is None:
         objects = gc.get_objects()
     if '.' in typename:
-        return sum(1 for o in objects if long_typename(o) == typename)
+        return sum(1 for o in objects if _long_typename(o) == typename)
     else:
         return sum(1 for o in objects if type(o).__name__ == typename)
 
@@ -116,13 +120,13 @@ def typestats(objects=None, shortnames=True):
     if objects is None:
         objects = gc.get_objects()
     if shortnames:
-        typename = short_typename
+        typename = _short_typename
     else:
-        typename = long_typename
+        typename = _long_typename
     stats = {}
-    for o in objects:
-        n = typename(o)
-        stats[n] = stats.get(n, 0) + 1
+    for obj in objects:
+        name = typename(obj)
+        stats[name] = stats.get(name, 0) + 1
     return stats
 
 
@@ -181,12 +185,12 @@ def show_most_common_types(limit=10, objects=None, shortnames=True):
 
     """
     stats = most_common_types(limit, objects, shortnames=shortnames)
-    width = max(len(name) for name, count in stats)
+    width = max(len(name) for name, _ in stats)
     for name, count in stats:
         print('%-*s %i' % (width, name, count))
 
 
-def show_growth(limit=10, peak_stats={}, shortnames=True):
+def show_growth(limit=10, peak_stats=None, shortnames=True):
     """Show the increase in peak object counts since last call.
 
     Limits the output to ``limit`` largest deltas.  You may set ``limit`` to
@@ -203,7 +207,7 @@ def show_growth(limit=10, peak_stats={}, shortnames=True):
         >>> objgraph.show_growth()
         wrapper_descriptor       970       +14
         tuple                  12282       +10
-        dict                    1922        +7
+        dict                    1922       +7
         ...
 
     .. versionadded:: 1.5
@@ -212,6 +216,8 @@ def show_growth(limit=10, peak_stats={}, shortnames=True):
        New parameter: ``shortnames``.
 
     """
+    if peak_stats is None:
+        peak_stats = {}
     gc.collect()
     stats = typestats(shortnames=shortnames)
     deltas = {}
@@ -220,8 +226,7 @@ def show_growth(limit=10, peak_stats={}, shortnames=True):
         if count > old_count:
             deltas[name] = count - old_count
             peak_stats[name] = count
-    deltas = sorted(deltas.items(), key=operator.itemgetter(1),
-                    reverse=True)
+    deltas = sorted(deltas.items(), key=operator.itemgetter(1), reverse=True)
     if limit:
         deltas = deltas[:limit]
     if deltas:
@@ -274,7 +279,7 @@ def by_type(typename, objects=None):
     if objects is None:
         objects = gc.get_objects()
     if '.' in typename:
-        return [o for o in objects if long_typename(o) == typename]
+        return [o for o in objects if _long_typename(o) == typename]
     else:
         return [o for o in objects if type(o).__name__ == typename]
 
@@ -408,15 +413,15 @@ def show_backrefs(objs, max_depth=3, extra_ignore=(), filter=None, too_many=10,
     .. versionchanged:: 1.8
        New parameter: ``shortnames``.
 
-    .. versionchanged:: 1.9
+    .. versionchanged:: 2.0
        New parameter: ``output``.
 
     """
     show_graph(objs, max_depth=max_depth, extra_ignore=extra_ignore,
                filter=filter, too_many=too_many, highlight=highlight,
                edge_func=gc.get_referrers, swap_source_target=False,
-               filename=filename, output=output, extra_info=extra_info, refcounts=refcounts,
-               shortnames=shortnames)
+               filename=filename, output=output, extra_info=extra_info,
+               refcounts=refcounts, shortnames=shortnames)
 
 
 def show_refs(objs, max_depth=3, extra_ignore=(), filter=None, too_many=10,
@@ -476,8 +481,9 @@ def show_refs(objs, max_depth=3, extra_ignore=(), filter=None, too_many=10,
     .. versionchanged:: 1.8
        New parameter: ``shortnames``.
 
-    .. versionchanged:: 1.9
+    .. versionchanged:: 2.0
        New parameter: ``output``.
+
     """
     show_graph(objs, max_depth=max_depth, extra_ignore=extra_ignore,
                filter=filter, too_many=too_many, highlight=highlight,
@@ -513,14 +519,15 @@ def show_chain(*chains, **kw):
     .. versionchanged:: 1.7
        New parameter: ``backrefs``.
 
-    .. versionchanged:: 1.9
+    .. versionchanged:: 2.0
        New parameter: ``output``.
 
     """
     backrefs = kw.pop('backrefs', True)
     chains = [chain for chain in chains if chain] # remove empty ones
-    def in_chains(x, ids=set(map(id, itertools.chain(*chains)))):
-        return id(x) in ids
+    ids = set(map(id, itertools.chain(*chains)))
+    def in_chains(obj):
+        return id(obj) in ids
     max_depth = max(map(len, chains)) - 1
     if backrefs:
         show_backrefs([chain[-1] for chain in chains], max_depth=max_depth,
@@ -555,6 +562,26 @@ def is_proper_module(obj):
 #
 
 def find_chain(obj, predicate, edge_func, max_depth=20, extra_ignore=()):
+    """Finds a single chain from the object until the predicate matches.
+
+    ``obj`` is the starting point of the chain.
+
+    ``predicate`` should be a function that takes an object and returns true
+    if the chain is complete.
+
+    `edge_func` is a function that given a node returns its neighbours.
+
+    `max_depth` is the maximum chain depth before the function should give up.
+    Defaults to 20.
+
+    `extra_ignore` an optional set of object ids that should be ignored
+    when computing neigbours.
+
+    Returns the chain or `[obj]` if no chain was found.
+
+    .. versionadded::1.8
+    """
+
     queue = [obj]
     depth = {id(obj): 0}
     parent = {id(obj): None}
@@ -564,8 +591,9 @@ def find_chain(obj, predicate, edge_func, max_depth=20, extra_ignore=()):
     ignore.add(id(depth))
     ignore.add(id(parent))
     ignore.add(id(ignore))
+    # pylint: disable=W0212
     ignore.add(id(sys._getframe()))  # this function
-    ignore.add(id(sys._getframe(1))) # find_chain/find_backref_chain, most likely
+    ignore.add(id(sys._getframe(1))) # find_chain/find_backref_chain
     gc.collect()
     while queue:
         target = queue.pop(0)
@@ -593,26 +621,124 @@ def show_graph(objs, edge_func, swap_source_target,
                max_depth=3, extra_ignore=(), filter=None, too_many=10,
                highlight=None, filename=None, extra_info=None,
                refcounts=False, shortnames=True, output=None):
-    if not isinstance(objs, (list, tuple)):
-        objs = [objs]
+    """Outputs a graph.
+
+    See :func:`show_refs` or :func:`show_backrefs` for more information.
+    """
+    dot_filename = None
     if filename and output:
         raise ValueError('Cannot specify both output and filename.')
-    elif output:
-        f = output
     elif filename and filename.endswith('.dot'):
-        f = codecs.open(filename, 'w', encoding='utf-8')
+        output = codecs.open(filename, 'w', encoding='utf-8')
         dot_filename = filename
-    else:
+    elif not output:
         fd, dot_filename = tempfile.mkstemp(prefix='objgraph-',
-                                            suffix='.dot', text=True)
-        f = os.fdopen(fd, "w")
-        if getattr(f, 'encoding', None):
+                                            suffix='.dot',
+                                            text=True)
+        output = os.fdopen(fd, 'w')
+        if getattr(output, 'encoding', None):
             # Python 3 will wrap the file in the user's preferred encoding
             # Re-wrap it for utf-8
             import io
-            f = io.TextIOWrapper(f.detach(), 'utf-8')
-    f.write('digraph ObjectGraph {\n'
-            '  node[shape=box, style=filled, fillcolor=white];\n')
+            output = io.TextIOWrapper(output.detach(), 'utf-8')
+
+    nodes = _build_dot(objs, edge_func, swap_source_target, output,
+                       max_depth=max_depth, extra_ignore=extra_ignore,
+                       filter=filter, too_many=too_many, highlight=highlight,
+                       extra_info=extra_info, refcounts=refcounts,
+                       shortnames=shortnames)
+    if not filename and not dot_filename:
+        return
+    # The file should only be closed if this function was in charge of opening
+    # the file.
+    output.close()
+    LOG.info('Graph written to %s (%d nodes)', dot_filename, nodes)
+    if filename and filename.endswith('.dot'):
+        # nothing else to do, the user asked for a .dot file
+        return
+    elif filename:
+        if _program_in_path('dot'):
+            _, ext = os.path.splitext(filename)
+            f = open(filename, 'wb')
+            dot = subprocess.Popen(['dot', ('-T' + ext[1:]), dot_filename],
+                                   stdout=f, close_fds=False)
+            dot.wait()
+            if dot.returncode != 0:
+                LOG.error('dot failed to generate "%s" image:'
+                          ' output format not supported?', filename)
+                raise ValueError('Output format not supported.')
+            f.close()
+            LOG.info('Image generated as %s', filename)
+        else:
+            LOG.error('Requesting a non-DOT file requires the dot program.'
+                      ' Outputing as file instead.')
+            raise ValueError('Output format not supported.')
+    else:
+        if _program_in_path('xdot'):
+            LOG.info('Spawning graph viewer (xdot)')
+            subprocess.Popen(['xdot', dot_filename], close_fds=True)
+        else:
+            LOG.error('Neither output nor filename specified, graph viewer '
+                      '(xdot) should be installed to display the graph.')
+            raise ValueError('Output format not supported.')
+
+
+def _output_node(output, obj, depth, max_depth, highlight,
+                 extra_info, refcounts, shortnames):
+    node_id = _obj_node_id(obj)
+    output.write('  %s[label="%s"];\n'
+                 % (node_id,
+                    _obj_label(obj, extra_info, refcounts, shortnames)))
+    hue, sat, val = _gradient((0, 0, 1), (0, 0, .3), depth, max_depth)
+    if depth == 0:
+        output.write('  %s[fontcolor=red];\n' % node_id)
+    if inspect.ismodule(obj):
+        hue = .3
+        sat = 1
+    if highlight and highlight(obj):
+        hue = .6
+        sat = .6
+        val = 0.5 + val * 0.5
+    output.write('  %s[fillcolor="%g,%g,%g"];\n'
+                 % (node_id, hue, sat, val))
+    if val < 0.5:
+        output.write('  %s[fontcolor=white];\n' % node_id)
+    if hasattr(getattr(obj, '__class__', None), '__del__'):
+        output.write('  %s->%s_has_a_del[color=red,style=dotted,len=0.25,'
+                     'weight=10];\n' %
+                     (node_id, node_id))
+        output.write('  %s_has_a_del[label="__del__",shape=doublecircle,'
+                     'height=0.25,color=red,fillcolor="0,.5,1",'
+                     'fontsize=6];\n'
+                     % node_id)
+
+def _output_skipped_node(output, obj, skipped, depth, max_depth,
+                         swap_source_target):
+    node_id = _obj_node_id(obj)
+    hue, sat, val = _gradient((0, 1, 1),
+                              (0, 1, .3),
+                              depth + 1, max_depth)
+    if swap_source_target:
+        label = '%d more references' % skipped
+        edge = '%s->too_many_%s' % (node_id, node_id)
+    else:
+        label = '%d more backreferences' % skipped
+        edge = 'too_many_%s->%s' % (node_id, node_id)
+    output.write('  %s[color=red,style=dotted,len=0.25,weight=10];\n'
+                 % edge)
+    output.write('  too_many_%s[label="%s",shape=box,height=0.25,'
+                 'color=red,fillcolor="%g,%g,%g",fontsize=6];\n'
+                 % (node_id, label, hue, sat, val))
+    output.write('  too_many_%s[fontcolor=white];\n' % node_id)
+
+def _build_dot(objs, edge_func, swap_source_target, output, max_depth=3,
+               extra_ignore=(), filter=None, too_many=10, highlight=None,
+               extra_info=None, refcounts=False, shortnames=True):
+    # Returns the number of nodes created.
+    if not isinstance(objs, (list, tuple)):
+        objs = [objs]
+    output.write('digraph ObjectGraph {\n'
+                 '  node[shape=box, style=filled, fillcolor=white];\n')
     queue = []
     depth = {}
     ignore = set(extra_ignore)
@@ -621,12 +747,14 @@ def show_graph(objs, edge_func, swap_source_target,
     ignore.add(id(queue))
     ignore.add(id(depth))
     ignore.add(id(ignore))
+    # pylint: disable=W0212
     ignore.add(id(sys._getframe()))  # this function
     ignore.add(id(sys._getframe().f_locals))
-    ignore.add(id(sys._getframe(1))) # show_refs/show_backrefs, most likely
+    ignore.add(id(sys._getframe(1))) # show_graph
     ignore.add(id(sys._getframe(1).f_locals))
+    ignore.add(id(sys._getframe(2))) # show_refs/show_backrefs, most likely
+    ignore.add(id(sys._getframe(2).f_locals))
     for obj in objs:
-        f.write('  %s[fontcolor=red];\n' % (obj_node_id(obj)))
         depth[id(obj)] = 0
         queue.append(obj)
         del obj
@@ -636,21 +764,8 @@ def show_graph(objs, edge_func, swap_source_target,
         nodes += 1
         target = queue.pop(0)
         tdepth = depth[id(target)]
-        f.write('  %s[label="%s"];\n' % (obj_node_id(target), obj_label(target, extra_info, refcounts, shortnames)))
-        h, s, v = gradient((0, 0, 1), (0, 0, .3), tdepth, max_depth)
-        if inspect.ismodule(target):
-            h = .3
-            s = 1
-        if highlight and highlight(target):
-            h = .6
-            s = .6
-            v = 0.5 + v * 0.5
-        f.write('  %s[fillcolor="%g,%g,%g"];\n' % (obj_node_id(target), h, s, v))
-        if v < 0.5:
-            f.write('  %s[fontcolor=white];\n' % (obj_node_id(target)))
-        if hasattr(getattr(target, '__class__', None), '__del__'):
-            f.write("  %s->%s_has_a_del[color=red,style=dotted,len=0.25,weight=10];\n" % (obj_node_id(target), obj_node_id(target)))
-            f.write('  %s_has_a_del[label="__del__",shape=doublecircle,height=0.25,color=red,fillcolor="0,.5,1",fontsize=6];\n' % (obj_node_id(target)))
+        _output_node(output, target, tdepth, max_depth, highlight,
+                     extra_info, refcounts, shortnames)
         if tdepth >= max_depth:
             continue
         if is_proper_module(target) and not swap_source_target:
@@ -675,8 +790,10 @@ def show_graph(objs, edge_func, swap_source_target,
                 srcnode, tgtnode = target, source
             else:
                 srcnode, tgtnode = source, target
-            elabel = edge_label(srcnode, tgtnode, shortnames)
-            f.write('  %s -> %s%s;\n' % (obj_node_id(srcnode), obj_node_id(tgtnode), elabel))
+            elabel = _edge_label(srcnode, tgtnode, shortnames)
+            output.write('  %s -> %s%s;\n' % (_obj_node_id(srcnode),
+                                              _obj_node_id(tgtnode),
+                                              elabel))
             if id(source) not in depth:
                 depth[id(source)] = tdepth + 1
                 queue.append(source)
@@ -684,84 +801,46 @@ def show_graph(objs, edge_func, swap_source_target,
             del source
         del neighbours
         if skipped > 0:
-            h, s, v = gradient((0, 1, 1), (0, 1, .3), tdepth + 1, max_depth)
-            if swap_source_target:
-                label = "%d more references" % skipped
-                edge = "%s->too_many_%s" % (obj_node_id(target), obj_node_id(target))
-            else:
-                label = "%d more backreferences" % skipped
-                edge = "too_many_%s->%s" % (obj_node_id(target), obj_node_id(target))
-            f.write('  %s[color=red,style=dotted,len=0.25,weight=10];\n' % edge)
-            f.write('  too_many_%s[label="%s",shape=box,height=0.25,color=red,fillcolor="%g,%g,%g",fontsize=6];\n' % (obj_node_id(target), label, h, s, v))
-            f.write('  too_many_%s[fontcolor=white];\n' % (obj_node_id(target)))
-    f.write("}\n")
-    if output:
-        return
-    # The file should only be closed if this function was in charge of opening
-    # the file.
-    f.close()
-    print("Graph written to %s (%d nodes)" % (dot_filename, nodes))
-    if filename and filename.endswith('.dot'):
-        # nothing else to do, the user asked for a .dot file
-        return
-    if not filename and program_in_path('xdot'):
-        print("Spawning graph viewer (xdot)")
-        subprocess.Popen(['xdot', dot_filename], close_fds=True)
-    elif program_in_path('dot'):
-        if not filename:
-            print("Graph viewer (xdot) not found, generating a png instead")
-            filename = dot_filename[:-4] + '.png'
-        stem, ext = os.path.splitext(filename)
-        f = open(filename, 'wb')
-        dot = subprocess.Popen(['dot', ('-T' + ext[1:]), dot_filename],
-                               stdout=f, close_fds=False)
-        dot.wait()
-        if dot.returncode != 0:
-            # XXX: shouldn't this go to stderr or a log?
-            print("dot failed to generate '%s' image: output format not supported?")
-        f.close()
-        print("Image generated as %s" % filename)
-    else:
-        if filename:
-            print("Graph viewer (xdot) and image renderer (dot) not found, not doing anything else")
-        else:
-            print("Unrecognized file type (%s), not doing anything else" % filename)
+            _output_skipped_node(output, target, skipped, tdepth, max_depth,
+                                 swap_source_target)
+    output.write('}\n')
+    return nodes
 
 
-def obj_node_id(obj):
+def _obj_node_id(obj):
     return ('o%d' % id(obj)).replace('-', '_')
 
 
-def obj_label(obj, extra_info=None, refcounts=False, shortnames=True):
+def _obj_label(obj, extra_info=None, refcounts=False, shortnames=True):
     if shortnames:
         label = [type(obj).__name__]
     else:
-        label = [long_typename(obj)]
+        label = [_long_typename(obj)]
     if refcounts:
         label[0] += ' [%d]' % (sys.getrefcount(obj) - 4)
         # Why -4?  To ignore the references coming from
-        #   obj_label's frame (obj)
+        #   _obj_label's frame (obj)
         #   show_graph's frame (target variable)
         #   sys.getrefcount()'s argument
         #   something else that doesn't show up in gc.get_referrers()
-    label.append(safe_repr(obj))
+    label.append(_safe_repr(obj))
     if extra_info:
         label.append(str(extra_info(obj)))
-    return quote('\n'.join(label))
+    return _quote('\n'.join(label))
 
 
-def quote(s):
-    return (s.replace("\\", "\\\\")
-             .replace("\"", "\\\"")
-             .replace("\n", "\\n")
-             .replace("\0", "\\\\0"))
+def _quote(s):
+    return (s.replace('\\', '\\\\')
+            .replace('\"', '\\\"')
+            .replace('\n', '\\n')
+            .replace('\0', '\\\\0'))
 
 
-def short_typename(obj):
+def _short_typename(obj):
     return type(obj).__name__
 
 
-def long_typename(obj):
+def _long_typename(obj):
     objtype = type(obj)
     name = objtype.__name__
     module = getattr(objtype, '__module__', None)
@@ -771,14 +850,15 @@ def long_typename(obj):
         return name
 
 
-def safe_repr(obj):
+def _safe_repr(obj):
+    # pylint: disable=W0702
     try:
-        return short_repr(obj)
+        return _short_repr(obj)
     except:
         return '(unrepresentable)'
 
 
-def short_repr(obj):
+def _short_repr(obj):
     if isinstance(obj, (type, types.ModuleType, types.BuiltinMethodType,
                         types.BuiltinFunctionType)):
         return obj.__name__
@@ -802,20 +882,20 @@ def short_repr(obj):
     return repr(obj)[:40]
 
 
-def gradient(start_color, end_color, depth, max_depth):
+def _gradient(start_color, end_color, depth, max_depth):
     if max_depth == 0:
         # avoid division by zero
         return start_color
-    h1, s1, v1 = start_color
-    h2, s2, v2 = end_color
-    f = float(depth) / max_depth
-    h = h1 * (1-f) + h2 * f
-    s = s1 * (1-f) + s2 * f
-    v = v1 * (1-f) + v2 * f
-    return h, s, v
+    hue1, sat1, val1 = start_color
+    hue2, sat2, val2 = end_color
+    fraction = float(depth) / max_depth
+    hue = hue1 * (1-fraction) + hue2 * fraction
+    sat = sat1 * (1-fraction) + sat2 * fraction
+    val = val1 * (1-fraction) + val2 * fraction
+    return hue, sat, val
 
 
-def edge_label(source, target, shortnames=True):
+def _edge_label(source, target, shortnames=True):
     if isinstance(target, dict) and target is getattr(source, '__dict__', None):
         return ' [label="__dict__",weight=10]'
     if isinstance(source, types.FrameType):
@@ -838,27 +918,27 @@ def edge_label(source, target, shortnames=True):
     if isinstance(source, types.FunctionType):
         for k in dir(source):
             if target is getattr(source, k):
-                return ' [label="%s",weight=10]' % quote(k)
+                return ' [label="%s",weight=10]' % _quote(k)
     if isinstance(source, dict):
         for k, v in iteritems(source):
             if v is target:
-                if isinstance(k, basestring) and is_identifier(k):
-                    return ' [label="%s",weight=2]' % quote(k)
+                if isinstance(k, basestring) and _is_identifier(k):
+                    return ' [label="%s",weight=2]' % _quote(k)
                 else:
                     if shortnames:
                         tn = type(k).__name__
                     else:
-                        tn = long_typename(k)
-                    return ' [label="%s"]' % quote(tn + "\n" + safe_repr(k))
+                        tn = _long_typename(k)
+                    return ' [label="%s"]' % _quote(tn + '\n' + _safe_repr(k))
     return ''
 
 
-is_identifier = re.compile('[a-zA-Z_][a-zA-Z_0-9]*$').match
+_is_identifier = re.compile('[a-zA-Z_][a-zA-Z_0-9]*$').match
 
 
-def program_in_path(program):
-    path = os.environ.get("PATH", os.defpath).split(os.pathsep)
-    path = [os.path.join(dir, program) for dir in path]
-    path = [True for file in path
-            if os.path.isfile(file) or os.path.isfile(file + '.exe')]
+def _program_in_path(program):
+    path = os.environ.get('PATH', os.defpath).split(os.pathsep)
+    path = [os.path.join(_dir, program) for _dir in path]
+    path = [True for _file in path
+            if os.path.isfile(_file) or os.path.isfile(_file + '.exe')]
     return bool(path)